1.Upgrade wxwidgets to 2.9.1(MS windows) 2.Patch for crash in rex Search 3.Patch for the but that open 2 files would crash under ubuntu 4.Patch for compiling under ubuntu, using wxwidgets 2.8

+++ src/MadEditFrame.cpp
@@ -3450,7 +3450,7 @@ void MadEditFrame::OnFilePageSetup(wxCommandEvent& event)
         wxString oldpath=m_Config->GetPath();
         m_Config->SetPath(wxT("/MadEdit"));
 
-        m_Config->Write(wxT("PageOrientation"), g_PageSetupData->GetPrintData().GetOrientation());
+        m_Config->Write(wxT("PageOrientation"), (long)g_PageSetupData->GetPrintData().GetOrientation());
 
         //((wxFrame*)wxTheApp->GetTopWindow())->SetTitle(wxString::Format(wxT("%d"), g_PageSetupData->GetPaperId()));
         wxSize size=g_PageSetupData->GetPaperSize();
